Move forwardspends to separate file
