+++ maediprojects/__init__.py
@@ -71,6 +71,7 @@ def register_blueprints(app):
     app.register_blueprint(views.management.blueprint)
     app.register_blueprint(views.counterpart_funding.blueprint)
     app.register_blueprint(views.activity_finances.blueprint)
+    app.register_blueprint(views.activity_forwardspends.blueprint)
 
 
 def register_errorhandlers(app):
