+++ maediprojects/views/__init__.py
@@ -1,4 +1,4 @@
 from maediprojects.views import activities, api, iati, \
 	codelists, users, exports, reports, \
 	management, counterpart_funding, \
-	activity_finances
+	activity_finances, activity_forwardspends
